+++ src/thermo/Phase.cpp
@@ -647,7 +647,7 @@ doublereal Phase::elementalMoleFraction(const size_t m) const
     checkElementIndex(m);
     doublereal Z_n = 0.0;
     for (size_t k = 0; k != m_kk; ++k) {
-        int nTotalAtoms = 0;
+        double nTotalAtoms = 0;
         for (size_t l = 0; l != m_mm; ++l) {
             nTotalAtoms += nAtoms(k, l);
         }
